分離完了
